Initial commit
For now, it simply reports and exits, without a loop.  The reason
is that the wmiirc script was trying to run something in a shell
while loop and pipe its output repeatedly through wmiir.  If this
program doesn't exit, that pipe doesn't get flushed, and no statusbar
is created.  I'll either have to inviestigate
	1) closing and re-opening stdout, which may not work anyway
	2) libixp from suckless.org.  Write to the 9p filesystem myself.